Move Ascii, Binary and ValueEncoding to its own file
That way it's more natural that both MetadataValue and MetadataKey use them.
