+++ src/lib.rs
@@ -22,6 +22,7 @@ pub mod generic;
 
 mod body;
 mod error;
+mod metadata_encoding;
 mod metadata_key;
 mod metadata_value;
 mod metadata_map;
@@ -38,9 +39,9 @@ pub use response::Response;
 /// The metadata module contains data structures and utilities for handling
 /// gRPC custom metadata.
 pub mod metadata {
-    pub use metadata_key::Ascii;
+    pub use metadata_encoding::Ascii;
+    pub use metadata_encoding::Binary;
     pub use metadata_key::AsciiMetadataKey;
-    pub use metadata_key::Binary;
     pub use metadata_key::BinaryMetadataKey;
     pub use metadata_key::MetadataKey;
     pub use metadata_value::MetadataValue;
